Rewrite Travis CI job as a bash script.
The way we were piping some of the commands in the bash pipeline inlined
in docker run command inlined in the the YAML file was unintentionally
silencing failues. This writes out the entire script for running Docker
commands as a bash script (with set -o pipefail), which makes it easier
to express and easier to avoid accidentally silencing output.

This also consolidates the "no internet" and "with internet" jobs
together, since the "no internet" one is stricter anyway.